Review feedbacks - query implementation merged with state store implementation, misc refactoring
