+++ core/src/main/resources/reference.conf
@@ -546,6 +546,9 @@ jdbc-read-journal {
 jdbc-durable-state-store {
   class = "akka.persistence.jdbc.state.scaladsl.DurableStateStore"
 
+  # New states are retrieved (polled) with this interval.
+  refreshInterval = "1s"
+
   tables {
     state {
       tableName = "state"
